+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - Dude color from message
 - Dude life time
+- Dude emotes spitter